Using toString method(if implemented) to log an object.
This improves performance since now one doesn't need
to convert an object to a string before logging it. The
object will be strigified using the toString method (if
implemented) just before being logged. So, if we log an
object with log-level "DEBUG" and the level of the logger
is "INFO", the object will never be stringified -- saving
valuable CPU cycles.