Add new appender logglyTagged
This appender allows adding tags to any call to the logger.

Documentation added:

Loggly Appender. Sends logging events to Loggly using node-loggly,
optionally setting tags.

This appender will scan the msg from the logging event, and pull out
any argument of the shape `{ tags: [] }` so that it's possible
to add tags in a normal logging call. If an object contains tags and
other members, that object is not used for loggly tagging.

For example:

logger.info({ tags: ['my-tag-1', 'my-tag-2'] }, 'Some message', someObj, ...)

And then this appender will remove the tags param and append it to the config.tags.