refactor: streams/*.js (test passed)
As class inheritance inside stream/*.js, standalone refactor don't work.
So streams/*.js are refactored in batch.
One thing worth mentioning is that, the old inherit-style will call the parent
constructor at the end of child constructor for some code initialization. However,
with ES6 class syntax, the call of `super` must be afore every statement with this.
So, some code initializing logic are reordered to suit the ES6 syntax and the tests.